Replace `id3` crate with `audiotags`
Note that the tests still use `id3` in the case when a tag is written to
a file to mock an audio file with a tag. This is because `audiotags`
doesn't support writing a tag to a file that doesn't contain an existing
tag.

Tests which involve writing a tag to a file as means of mocking an audio
file are thus limited to only being able to mock mp3 files, but this is
a tolerable limitation.